Move user_badge to forum module
