+++ oioioi/forum/templates/forum/category.html
@@ -3,6 +3,7 @@
 {% load check_perm %}
 {% load get_user_name %}
 {% load pagination_tags %}
+{% load user_badge %}
 
 {% block title %}{% trans "Category" %}: {{ category.name }}{% endblock %}
